libnmstate: Support bond slaves configuration
Slaves can now be defined for a bond.

Note that in the nmstate schema, slaves are specified under the bond
interface while NM expects each slave to be marked as "slave".
In order to adjust, the desired state is normalized with slaves metadata
in the form of an internal property "master" which takes the value of
the bond master.