libnmstate: Include master type in the metadata
In addition to the master iface name, this patch adds the type of the
master, in order to use this information when creating the slave
profile (NM).

The metadata update has been generalized in order to allow each
master type to include slave specific properties. See OVS bridge usage
in the following patch.