libnmstate: Support ovs-bridge configuration
It is now possible to define an OVS bride with basic properties and
attach to it system interfaces (as ports).

Note that for attaching a system iface, NM requires a port profile to be
defined which connect the iface to the bridge. NMState implicitly
creates these port profiles but does not report or require one to define
them.

The declerative configuration is such a bridge looks like this:
{
    "interfaces": [
        {
            "bridge": {
                "options": {
                    "mcast-snooping-enable": false,
                    "rstp": false,
                    "stp": true
                },
                "port": [
                    {
                        "name": "eth3",
                        "type": "system"
                    }
                ]
            },
            "ipv4": {
                "enabled": false
            },
            "name": "ovs-br0",
            "state": "up",
            "type": "ovs-bridge"
        }
    ]
}