libnmstate: Use a main-loop for the NM provider
The NM provider is executing libnm async funtions.
For proper operation and in order to serialize the actions, GLib
main-loop use has been embedded through a mainloop object.

The async funtions against libnm are registered to the mainloop object
and then executed one after the other from that queue.
The callback functions for each async command are responsible to execute
the next action in the mainloop object queue.

Signed-off-by: Edward Haas <edwardh@redhat.com>