nm/checkpoint: Introduce transaction capabilities
The NM checkpoint option enables nmstate to "save" the state of the
configuration prior to the change apply, and in case any failure is
detected/raised, it can restore/rollback to that saved state.

This patch is aimed to handle rollbacks in case of errors during the
setup.

Note: The NM checkpoint has been accessed through d-bus directly and not
through libnm. libnm support for checkpoint is available from version 1.12,
making it difficult to test it on Centos 7.5 and Fedora 28.

Signed-off-by: Edward Haas <edwardh@redhat.com>