netapplier: Reflect correctly slave master even when reused
In a scenario where an existing slave is specified under a new master in
the desired state, but not removed from the current master, the outcome
desired state (through the metadata) should point to the new master.

This patch fixes a bug where such slaves pointed to the previous master
in the calculated desired state.

Signed-off-by: Edward Haas <edwardh@redhat.com>