nm/applier: OVS slaves need special admin state handling
For OVS birdges, the admin state handler needs to process all the bridge
slaves (ports and interface devices) by changing their admin state as
well.
The order is also important, especially when activating the devices: The
port ones need to be activated first and only then the interfaces.
The scenario in which this has been found to matter: Moving a slave from
one birdge to another (without removing it from the old one).

Signed-off-by: Edward Haas <edwardh@redhat.com>