libnmstate: Refresh current state before the edit step
The mainloop is run in two steps, once for addition and a second time
for editing.
The split is introduced because after the addition, the current state
needs to be refreshed and be based on the addition step. This implies
that the addition should finish completely (running all the actions).

It is now possible to create a new master (ovs-bridge) to use a slave
that is already used by a different master. The slave will "move" to the
new master correctly.

Signed-off-by: Edward Haas <edwardh@redhat.com>