netapplier: Add a sleep workaround to support verify correctly
When verify is enabled, nmstate reads the current state and compares to
the desired one, failing if they are not equal.
The current state is read imidiatly after the apply of all ifaces is
completed, but this seems to be too soon, as some devices are still in
process of reaching activated state but not there yet.

A proper solution will be provided in the near future to wait for the
proper state of each device before proceeding.

Signed-off-by: Edward Haas <edwardh@redhat.com>