automation: Connect 2 additional networks to docker
Adding 2 additional links to the running docker (3 in total) to allow
complex test configurations to be supported.

Due to the way docker adds the two additional interfaces (while the
container is already running with NetworkManager, docker is editing the
IP configuration), NetworkManager is stopped before the network is added
and started again after. This avoids the scenario where NM detects the
IP changes on the interface and therefore generates an in-memory
profile.

Signed-off-by: Edward Haas <edwardh@redhat.com>