nm: Expose device connection profile only if active
Fetching a device connection profile based on the device should only
consider activated connection profiles and not in transient ones.

The caller code is usually not expecting the returned connection profile
to be in between states. In a transient state, the connection profile
may miss several expected properties, like the interfaces it is bound
to.

Signed-off-by: Edward Haas <edwardh@redhat.com>