nm: Deactivate device before deleting its connection
Currently, on iface admin state change to 'down' or 'absent', nmstate is
issuing a connection delete command.
This is insufficient as NM will attempt to create a new default
connection profile for the device (unless specifically requested no to
in NM config file).

To instruct NM not to activate a default connection profile, the device
is now deactivated prior to deleting the connection profile.

With this change, it should be possible to remove the 'no-auto-default'
settings in the NM config (introduced in a following patch).

Signed-off-by: Edward Haas <edwardh@redhat.com>