nm: Execute all actions from within the mainloop
The execution of the first mainloop action has been executed outside the
mainloop, expecting it to trigger a callback that will  continue
from within the mainloop.
This flow fails when only a single action exists and it does not trigger
a callback, but finishes with a mainloop quit. As the mainloop has not
even started and there are no callbacks, it will just timeout.

This flow is now changed by executing all registered actions,
including the first one from within the mainloop.

This change directly resolved a bug in which an iface in 'down' state
could not be set again to 'down'.

Signed-off-by: Edward Haas <edwardh@redhat.com>