nm: Set a connection to autoconnect=false on creation or editing
When running multiple integration tests, nmstate randomly fails when
interacting with NM.

Logs and further debugging revealed that when a connection profile is
created, it automatically triggers an activation sequence on the
connection. When the nmsatate explicit `activate` command is executed,
it cancels the previous auto one and tries to execute an `activate`
command which on CI it has been found to fail frequently.

Until this is handled (in a following patch) the `autoconnect` property
is set to False. As a result, on reboot the connection profiles won't
get activated.

Signed-off-by: Edward Haas <edwardh@redhat.com>