netapplier: Remove absent ifaces from dstate on verification
When an iface state is set to `absent` in the desired state, the
verification step now ignores the entry when comparing with the
current state (as in the current state it should be missing).

There are interfaces which cannot be removed (e.g. physical nic/s) and
therefore the `absent` state is not valid. However, nmstate does not
distinguish between `down` and `absent` admin states, nor it knows to
distinguish between the removable and non-removable interfaces.
For the time being, this limitation is logged as an xfail test.

Signed-off-by: Edward Haas <edwardh@redhat.com>