nm ipv6: Enable IPv6 even without explicit addresses
IPv6 is now enabled even when the desired state includes IPv6
configuration enabled but without addresses or with link-local addresses only.

The IPv6 method is set to link-local in such cases.

This change also includes adjustments to the integration tests. IPv6 is
expected to be enabled only if specified explicitly.

Signed-off-by: Edward Haas <edwardh@redhat.com>