netapplier: Fix OVS proxy port handling in the edit step
The OVS port interfaces are created internally but not exposed to the
user at the nmstate API. In the edit step, these interfaces have not been
passed to the admin state function handler for processing.

The introduction of autoconnect=True cause this problem to be a stopper,
failing most of the tests.

The OVS port interfaces are now processed by the admin state handler,
causing a regular activation of OVS port interfaces.

Signed-off-by: Edward Haas <edwardh@redhat.com>