nm ovs: Fix slave interface detection
Given a master interface, slave interfaces are detected by checking
their `master` property.

The existing implementation checks the relation based on the master
device object instance (address), which seems not to work correctly
in all scenarios (probably dependent on the flow which calls the helper
function).
To resolve this issue, the relationship is checked based on the
interface name which is extracted from the device object.

Signed-off-by: Edward Haas <edwardh@redhat.com>