nmclient: Explicitly create a new client object only in netinfo
Usage of the NM client object should be limited in most cases to a
single instance.
In particular, during a mainloop run, only a single instance of the
client should be used due to its cached data.

This change is limited in scope, fixing issues seen with current
integration tests. A follow up change will organize the client object
lifetime in a more systematic manner.

Signed-off-by: Edward Haas <edwardh@redhat.com>