nm connection: Protect ConnectionProfile usage
Protect the ConnectionProfile object by limiting the ability to
re-import or re-create a profile once the object state represents one.

The object may be set with a device and/or a connection ID, prior to an
import action, allowing the object to be partially set with the
indentifiers and get loaded later on in the flow.
It is particularly useful when the device or connection ID is known in
advance, but the existance of the NM entity is available later in the
async flow.

The ability to reset the ConnectionProfile backend profile state is also
introduced, allowing the async retry mechinism to attempt and
load/import the profile into the object.

Signed-off-by: Edward Haas <edwardh@redhat.com>