integ tests: Skip route versification while testing bonds
The bond rollback test is set to create a bond successfully and then
fail at the verification step on an unsupported parameter, triggering a
rollback to the previous state (removing the bond).

The test first reads the current state, applies the desired state (which
failes with a rollbacl) and then checks that the new current state is equal
to the old current state.

With the addition of the routes entries, this test now fails frequently
on the default IPv6 route metric change [1].

1. https://nmstate.atlassian.net/browse/NMSTATE-162

Signed-off-by: Edward Haas <edwardh@redhat.com>