mac-address: Do not set the mac if identical to the existing
It has been observed that when always setting the mac address (even when
it is no different from the existing one) tests fail frequently.

This change introduces a wrokaround by not setting the mac address if it
is no different from the current one.

Hopefully this change will not get merged.

Signed-off-by: Edward Haas <edwardh@redhat.com>