nm.device: Introduce reapply functionality
Nmstate NM provider has been applying changes to interfaces by creating or
editing connection profiles and then activating them.
The activation operation is fully redefining the interface which may
cause traffic disruption (if traffic is passing through).

NM provides an alternative method for applying changes, named 'reapply'.
Changes to interfaces which are not expected to cause traffic disruption
can be applied using 'reapply'. If the change is not supported by
'reapply', it will fail the operation.

Signed-off-by: Edward Haas <edwardh@redhat.com>