Travis: Use flags for run-tests.sh
Use flags to configure run-tests.sh instead of internal variables. This
ensures that the public interface is used and makes it easier to use
other flags (such as `--copr`), as well. The docker image needs to be
specified using the internal variable because it is also used in
`before_install`.

Signed-off-by: Till Maas <opensource@till.name>