automation: Install Coveralls only for user
python-coveralls requires urllib3 which is also packaged and used by
yum-plugin-copr (as Python 2 package). Install coveralls with for
Python 3.6 and into the user space to avoid any conflicts with the RPM
package. The additional advantage is that this will also ensure that any
Python 2 incompatibilities that might occur in the future in
python-coveralls will not matter. This also separates the yum from the
pip installation steps to cleanup the container spec and allow to build
the container image with caching when developing/testing locally.

Signed-off-by: Till Maas <opensource@till.name>