nm.device: Introduce the modify operation
NM is exposing two operations to apply a profile: `activate` and `reapply`.
The `modify` operation is combining the two, exposing an abstraction in
which the knowledge of which backend operation to use is hidden.

`modify` implements a fallback from `reapply` to `activate`.

In case of special parameters changes (mtu), where the reapply operation
is not supported but fails silently, or when the reapply operation fails
explicitly, fallback to a connection activation.

Signed-off-by: Edward Haas <edwardh@redhat.com>