automation: Fix Fedora29 dev image
Udev has stopped working in the Fedora dev image.
It seems that the source is related to a missing systemd-udev package.
systemd-udev content has been originally (in CentOS) part of the systemd
package and in Fedora it has been extracted to its own package.
Until recently, the systemd-udev package has been required by the dracut
package and therefore present on the base image. However, the dracut package
has been dropped from the base image and therefore also its dependency,
systemd-udev is no longer present with the base Fedora image.

https://bugzilla.redhat.com/1704456 has been opened to track if
NetworkManager should directly depend on the systemd-udev package.

The container definition has been updated to include systemd-udev.

systemd-udevd is also being started when the container is run and NM is
restarted in order to confirm it loads with udev activated already.

Signed-off-by: Edward Haas <edwardh@redhat.com>