state: Generalize and optimize RouteEntry
The RouteEntry object represents a route entry.
Adjust RouteEntry to become a bit more generic, decoupling it from the
knowledge on how it is used.

The route state (absent) is now consistently not taking part of any equality,
hashing or matching operations. It is left to the caller to determine
which entries to operate on.
The implementation of `__lt__` enables sorting logic to be embedded in
the object.
To support an easy detection of absent routes, a helper property has
been added (`absent`).

Signed-off-by: Edward Haas <edwardh@redhat.com>