automation: Keep project folder clean during tests
Build and test artifacts are avoided by copying the project to a
temporary folder inside the container. Leaving the original folder
clean.

An optional `artifacts-dir` argument is added to control the
exported-artifacts folder location on the host. To keep the source
folder clean, locate the exported-artifacts folder outside the source folder.
Example: `--artifacts-dir /tmp/`

Signed-off-by: Edward Haas <edwardh@redhat.com>