tests integ: Adding helper and assert functions for MAC
Adding get_all_macs(state) helper function that returns the
MAC-address list of all the interfaces in the state. Since
the function call recreates multiple lists just to fetch one
item out of them, an assert_all_equal(items) is called
which asserts if all the MAC addresses are the same or not.

Signed-off-by: Nandan Kulkarni <nakulkar@redhat.com>