netapplier: Initialize mainloop global object on failure
The mainloop object is a singletone.
When libnmstate is used by an application which calls it multiple times
(e.g. running as a daemon), on mainloop (run) failure, the instance is kept
and reused in the following calls. Unfortunately, when the mainloop run fails,
its state may no longer be usable and may require initialization.

Therefore, to support such usages, the mainloop singletone is initialize
in case a mainloop failure is detected.

Signed-off-by: Edward Haas <edwardh@redhat.com>