netapplier: Unify the two mainloop runs into one
On apply, two mainloop runs are currently executed: One for adding new
interfaces and another for editing existing ones.
The original reasoning for the separation was the assumption that once
the new interfaces are created, there is a need to refresh the current
state with the new information, serving the editation step.

It seems that the current state refresh is not needed, at least based on
the current test coverage.
Therefore, this change unifies the two mainloop runs into a single one.

Signed-off-by: Edward Haas <edwardh@redhat.com>