schema: add linux bridge vlan filtering
The defined model allows for setting trunk and access ports.
Furthermore, it allows defining a native vlan on a trunk port.

It is possible to independently define the trunks/access vlans per
each port, and while NM / linux bridges allows it, at the moment,
we will not allow setting allowed vlans on the bridge port.
It will be hardcoded with the linux bridge default: having vlan id
1 as native.
Defining the schema of the uplink port will be done in a separate
PR.

This patch adds to the model:
  - a 'vlan' map to the bridge ports, that will hold a
    list trunk tag (that can be individual tags, or tag ranges), an
    access tag - to be used for access ports - and an 'enable-native'
    flag, intended to select one of the trunk tags as native vlan.

An example of a bridge configuration that allows for all vlans on a
trunk port:
- name: linux-br0
  type: linux-bridge
  state: up
  bridge:
    options:
      stp:
        enabled: true
    port:
      - name: eth1
        vlans:
          type: trunk
          trunk-tags:
            - id-range:
                min: 1
                max: 4095

An example of a bridge configuration with an access port:
- name: linux-br0
  type: linux-bridge
  state: up
  bridge:
    options:
      stp:
        enabled: true
    port:
      - name: eth1
        vlan:
          type: access
          tag: 101

An example of a bridge configuration having a port with a native vlan
trunk tag:
- name: linux-br0
  type: linux-bridge
  state: up
  bridge:
    options:
      stp:
        enabled: true
    port:
      - name: eth1
        vlan:
          type: trunk
          trunk-tags:
            - id: 101
            - id-range:
                min: 200
                max: 4095
          tag: 100
          enable-native: true

The bug that tracks the feature in NetworkManager can be found
at [0].

More information on this feature can be found in [1].

[0] - https://bugzilla.redhat.com/show_bug.cgi?id=1652910#c11
[1] - https://developers.redhat.com/blog/2017/09/14/vlan-filter-support-on-bridge/

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>