ipv6: Disable the IPv6 stack when explicitly requested
NetworkManager is currently not supporting the ability to disable the
IPv6 stack on the kernel [1].

As an intermediate solution, sysfs is used to control the IPv6 stack.

Rollback support is currently limited, not supporting a rollback from
an IPv6 disabled stack to an IPv6 enabled stack. Relevant tests have
been included.

[1] https://bugzilla.redhat.com/1643841

Signed-off-by: Edward Haas <edwardh@redhat.com>