nm.applier: Do not remove slaves that are mentioned with state UP
In scenarios where a master interface (bond, bridge) is removed but at
the same time one of the slaves is specified with a state UP, the slave
should not be removed.
The existing NM provider implementation is removing all slaves under the
removed master, regardless of the explicit desired admin state of the slaves
interfaces.

Adjust the logic to take into account the explicit desired state for the
interfaces when they get classified for removal. An explicit admin state
mentioned in the desired state should take priority over its removal.

Signed-off-by: Edward Haas <edwardh@redhat.com>