tests, integ: Extract iface-up fixtures out of conftest
The existing eth{1,2}_up fixtures serve the integration tests with the
majority of tests. There are however serveral tests which use unique
interfaces (e.g. dhcp tests) which need similar fixtures.

In order to share the functionality, the fixtures implementation have
been extracted to a shared location (testlib) and are aimed to be used
in the following patches by other more focused scoped fixtures.

Signed-off-by: Edward Haas <edwardh@redhat.com>