tests, integ: Fix dynamic IPv6 related tests
The dynamic IPv6 tests assumed wrongly that the IPv6 stack is always
enabled and have not reset the client interface config on each test.
This change adds explicit IPv6 enablement for the relevant interfaces
and introduces a fixture that provides a clean interface to the tests.

In addition, the existing xfail test has been relocated to the tail as
it was found to remove the IPv6 config from the server interface (on
which the dhcp/ra server is running).

Signed-off-by: Edward Haas <edwardh@redhat.com>