tests integ: Explicitly enable the dhcpcli iface with IP auto
The dhcpcli interface in the dhcp_tests module setup is consuming the
dynamic IP services.
The fixure that represented the iface just set it with state=UP on
setup and with state=DOWN on teardown.
The assumption was that without specifying any other details, NM will
define its profile with IPv4/6 auto.

The assumption does not work well with IPv6, therefore this change is
explicitly setting its state with DHCPv4/6 and Autoconf enabled.

Signed-off-by: Edward Haas <edwardh@redhat.com>