nm: Fix race problem on slave been deactivated by master
When master is activating, the `AutoconnectSlaves:Yes` will
activate the slaves, this will deactivate slave first.
There is a small chance(less than 1/1000), the
`waitfor_active_connection_async()` fails as the cached active
connection is disconnected.

The root cause is the active connection been replaced and marked as
DEACTIVATED with reason device disconnected by the following activate calls.

The fix is to get current active connection from NM.Device, if that is
not the cached one, use the new one for tracking 'state-change' signals.

Signed-off-by: Gris Ge <fge@redhat.com>