tests, integ: dynamic IP addresses hanging on vlans
On more complex scenarios, involving VLANs, NetworkManager
hangs while assigning an IP address via DHCP to the linux bridge.

This is examplified when creating a VLAN on top of a bond, on top
of an ethernet interface. That VLAN is in turn set as a port to a
linux bridge.

The 2 added tests show that NM hangs when the linux bridge has a
dynamic IP configuration on top of a VLAN - in the first test, this
happens within a single transaction, while on the last, it manifests
when multiple transactions are used.

The NM main loop aborts since the interfaces are not assigned a
dynamic IP address - since the DHCP server configured in the
dynamic IP module is not being ran behind a VLAN interface, and as
such, the L2 traffic (untagged) is discarded before reaching the
bridge.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>