state: allow empty interface description
If the description of an interface is set, the way to remove it is by
setting an empty description. Currently, this is not possible if we
activate the verification.

When the verification is performed, the "description" field appears
empty in the desired state, however in the current state the
"description" field does not appear so they do not match. This causes an
error.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>