GOVERNANCE: Introduce a meritocratic governance document
As the project takes momentum and members, users and contributors alike, are
increasing gradually, there is a need to document the project
governance.

The document is aimed to clarify and formalize the project different
member roles and the decision making model.

Nmstate is a meritocratic, consensus-based community project.
The document is based on the OSS Watch meritocratic governance
document template [1] with simplification changes.

[1] http://oss-watch.ac.uk/resources/meritocraticgovernancemodel

Signed-off-by: Edward Haas <edwardh@redhat.com>