nm bond: Handle bond get_info() while deleting
The test `test_bond_with_a_slave` of `nm/bond_test.py` will fail on

    AssertionError: assert not {'options': {}, 'slaves': []}

Due to the race of doing nm/bond.py get_info() against newly deleted
bond interface.

In order to fix the above scenario, treat the data as if the interface
does not exist and return `{}`.

Signed-off-by: Gris Ge <fge@redhat.com>