nm.device: support reapply on MTU changes
Support reapply on MTU changes as it is supported by NetworkManager
again. In addition, we are supporting the workaround if the
NetworkManager version is lower than 1.18.

Please note that test_change_mtu_with_stable_link_up integration test is
passing because there is no link DOWN or UNKNOWN events on monitoring.

Ref: https://bugzilla.redhat.com/1702657

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>