automation: Enhance pytest header
Mention NetworkManager RPM and Operating System in pytest header.
Example:

```
================= test session starts ==================
[...]
RPMs: NetworkManager-1.20.5-1.cab2e4fc96.el8.x86_64
OS: CentOS Linux 8 (Core)
```

This makes it easier to identify the test environment from the pytest
log. For example when the tests are running on bare-metal or in a VM
instead of the container that is setup with run-tests.sh.

Signed-off-by: Till Maas <opensource@till.name>