NM, applier: Explicitly remove any virtual device
With exception of veth interfaces that are currently
excluded until there is proper solution how to avoid
using veth as main interfaces in container.

If the action fails it is asssumed the operation is already complete
or in progress.

This is explicitly required due to two main reasons:
- Devices that are created without NM but later managed
by NM will not get delete when the profiles are removed.
- There seem to be an issue that devices are not removed
even when the device is created by NM.

Signed-off-by: Ales Musil <amusil@redhat.com>