Schema: Deprecate old LinuxBridge PORT_ constants
Deprecate the `LinuxBridge.PORT_*` constants that moved to
`LinuxBridge.Port` by using a metaclass that patches `__getattribute__`
to warn about the deprecation and take the value from the new constant.
The type is configured by the DEPRECATED_CONSTANTS dictionary that
contains mappings from the deprecated constant names to the new constant
names.

Signed-off-by: Till Maas <opensource@till.name>