nm.device: Deletion: Assume removed interface as success
Instead of making various and wrong assumption in
`_delete_device_callback()` which errors mean that the interface is
removed, just check if the interface is not there anymore. This will
also handle failures like `nm-device-error-quark: This device is not a
software device or is not realized (5)` or `error=nm-client-error-quark:
Object is no longer in the client cache (0)` that were not caught by the
previous code. When these errors happened or probably when
`NMDevice.is_real()` returns `False`, `NMDevice.get_iface()` returns
`None`. Therefore get the interface earlier to have it available in the
callback. A better solution might be to refactor Nmstate to properly
pass a context about which part of the desired state is handled in each
async call to ensure the interface name and possibly other useful
information is always present.

Signed-off-by: Till Maas <opensource@till.name>