nm.checkpoint: Initialize the dbus connection per instance
The checkpoint module is using dbus directly.
The current implementation uses a global dbus connection, however, if
NM service is restarted, this connection needs to be refreshed.

As the checkpoint is the only client of the dbus connection, the dbus
connection is now initialized on each instantiation of the checkpoint
object.

Signed-off-by: Edward Haas <edwardh@redhat.com>