state: Fix bridge type interface state parsing
A bridge type interface state includes the `bridge` subconfig only if it
is in an UP state.
When a bridge exists on the system but has no NM profile, it will appear
with state DOWN and will miss the bridge subconfig tree.

When parsing the current bridge interfaces, ignore the ones that their
state is not UP.

Signed-off-by: Edward Haas <edwardh@redhat.com>