Container: Specify registry for container images
To avoid name squatting attacks/problems it is necessary to specify the
container registry when specifying a container image. Otherwise an image
with the same name from a different registry could be obtained. This
needs to be fixed in `run-tests.sh` and `.travis.yml` as well but needs
special care to ensure that rebuilding containers locally still works.

References: https://raesene.github.io/blog/2019/09/25/typosquatting-in-a-multi-registry-world/

Signed-off-by: Till Maas <opensource@till.name>