GitHub: Add stale probot configuration.
To ensure that old/untouched pull requests do not pile up, configure the
probot stale bot to mark pull requests without activity in the last 14
days as stale and close them after 14 more days if there is no further
activity. The pull requests are marked with the `stale` label.

There are two possible solutions for this problem, the stale probot,
that is used here and the GitHub stale action. The stale probot seems to
unmark pull requests once there is activity while the GitHub stale
action requires the label to be removed manually. Since the latter means
more work, use the probot solution.

Signed-off-by: Till Maas <opensource@till.name>