nm, applier, ovs: Hash the OVS port device name
The OVS port device name is limited by NM to the kernel interface name
length (15 chars) even though there is no real underline interface. [1]

As a workaround, this change hashes the profile name into a max of 15 hex
characters and uses it as the profile interface/device name.
The interface name is creaded as part of the OVS port (proxy) interface
state and passed on through the setup flow by adding it as metadata to
the state.

[1] https://bugzilla.redhat.com/1788432

Signed-off-by: Gris Ge <fge@redhat.com>
Signed-off-by: Edward Haas <edwardh@redhat.com>