validator: Validate duplicate interface names
The current interface state is using the name property as an unique
key. Passing in duplicate interfaces names causes a failure only at the
verification step (if it is executed), making it hard to understand the
problem.

This change adds a validation for duplicate interface names, raising an
error early on.

Signed-off-by: Edward Haas <edwardh@redhat.com>