nm, ovs: Remove translate port options step
The OVS port options translation step has no added value at the moment.
Originally, the intention was to use a translation step between the nmstate
public state and the nm provider one, assuming that the two do not have
to be the same. However, the translation did not evolve to anything
useful and an 1:1 relation was always kept.

Therefore, it is dropped.

Signed-off-by: Edward Haas <edwardh@redhat.com>