nm, ovs: Remove translate bridge options step
The OVS bridge options translation step is translating a dict to a dict
with the same key & value to the same key & value.
Originally, the intention was to use a translation step between the nmstate
public state and the nm provider one, assuming that the two do not have
to be the same. However, the translation did not evolve to anything
useful and an 1:1 relation was always kept.

This has no added value at the moment and therefore it is dropped.

Signed-off-by: Edward Haas <edwardh@redhat.com>