nm, ovs: Support link-aggregation port
Link aggregation is implemented in the nm.ovs module, supporting both
setting and reporting the state.

The state (report) is interpreted from the NM configuration.

As part of the change, existing bond related properties (updelay and
downdelay) have been relocated to the link-aggregation subtree.

The lacp property has been integrated into the `mode` property.
Specifically, lacp is set to `active` when the lag mode is "lacp" or
"balance-tcp" and to "off" when the lag mode is "active-backup" or
"balance-slb".

Signed-off-by: Edward Haas <edwardh@redhat.com>