libnmstate: Do not remove slaves implicitly with master
When a master interface is removed, nmstate should not implicitly remove
its slaves. If such an action is desired, the user needs to explicitly
specify it.

This change scope includes the following interface types:
- Bond
- Linux Bridge
- OVS Bridge
- Team

Note: OVS internal interfaces (ovs-interface type) are an exception
here, as such interfaces may not exist without an OVS bridge. Therefore,
these are expected to get removed together with their bridge.

Signed-off-by: Edward Haas <edwardh@redhat.com>