automation: Fix container leftover on script exit
When the run-test.sh script exits due to an error the running container
is not always removed.
The reason has to do with an attempt to remove `*nmstate*.rpm` files
that fails and therefore does not continue to the next command.

This change makes sure the container removal is executed even if the
rpm file removal fails.

Signed-off-by: Edward Haas <edwardh@redhat.com>