nm, bond: Report only non-default settings
Except the `mode` bond option, report only the non-default option
values.

This change is avoiding an undesired merge of default values to the
desired state, which in turn causes defaults to be explicitly set on the
NM profile. Setting some defaults explicitly may cause validation errors
that are introduced by NM.

Signed-off-by: Edward Haas <edwardh@redhat.com>