validator: Check IP config in slaves or ports.
When an interface is created, the slaves/ports cannot have their IP
configuration enabled. In order to fix this problem, this patch
introduces a validation check to see whether the slave interfaces had
their IP enabled and if so, it raises an NmstateValueError.

There is an exception for OVS-Interface, which can have its
IP configuration enabled when set as a port.

Ref: https://bugzilla.redhat.com/1782674

Signed-off-by: Elvira García Ruiz <elviragr@riseup.net>