linux bridge: Merge port subtree
When merging bridge iface state, also perform merge of the bridge
ports.

This extension of the interface merging mechanism simplifies the
user's life in scenarios where the user wants to partially update
the bridge's configuration - e.g. update only the bridge's port
STP configuration - since it enables the user to simply specify
the updated values rather than the full current state + the updated
value.

Considering the current state (in yaml format):
---
interfaces:
  - name: eth1
    type: ethernet
    state: up
  - name: linux-br0
    type: linux-bridge
    state: up
    bridge:
      port:
        - name: eth1
          stp-hairpin-mode: false
          stp-path-cost: 100
          stp-priority: 32
          vlan:
            mode: trunk
            trunk-tags:
            - id: 101
            - id-range:
                min: 500
                max: 599
            tag: 100
            enable-native: true

If the user wants to remove the port vlan configuration on port
eth1 - and as a result disable port vlan filtering on the whole
bridge - the following yaml should be used:
---
interfaces:
  - name: linux-br0
    type: linux-bridge
    state: up
    bridge:
      port:
        - name: eth1
          vlan: {}

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>