linux bridge, vlan filtering: Normalize linux bridge ports
This normalization of the linux bridge port vlan subtree adds
an empty vlan config subtree from the port *when* it is empty.

By doing so, the states can be considered equal when the vlan
subtree is not present, and when the vlan subtree is empty. This
is required for the scenario where the vlan configuration is
removed from a port.

This normalization is used in the verification step.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>