schema, ovs: introduce OVS Patch support
This patch introduces the schema for OVS Patch support. In order to add
a patch interface to a OVS bridge, the user needs to define an OVS
interface with the patch configuration in the proper subtree.

In addition, it introduces the implementation needed for the OVS patch
port support.

Example:

```
interfaces:
- name: patch0
  type: ovs-interface
  state: up
  patch:
    peer: patch1
- name: ovs-br0
  type: ovs-bridge
  state: up
  bridge:
    options:
      stp: true
    port:
    - name: patch0
- name: patch1
  type: ovs-interface
  state: up
  patch:
    peer: patch0
- name: ovs-br1
  type: ovs-bridge
  state: up
  bridge:
    options:
      stp: true
    port:
    - name: patch1
```
Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>