libnmstate: Implement non persistent state support
This patch is introducing non persisntent state support by the keyword argument
save_to_disk when using libnmstate API apply() and edit() operations. By
the default save_to_disk is True.

All the changes done to the interfaces when using non persistent state
will be wiped out after rebooting the system.

In addition, this patch is introducing nmstatectl --memory-only support.
By the default, it is disabled so the argument is needed in order to use
non persistent states.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>