nm/connection: Wait on rollback
On slow server[1], test case `test_disable_static_ipv6_and_rollback` is
constant failing.

The root cause:
    The `profile.import_by_device(nm_dev)` does not actually import the
    profile as there is no `NM.ActiveConnection` at that time.
    Hence the follow up `profile.wait_dev_activation()` does nothing as
    `profile._nm_dev` is None.

To fix that, don't use `import_by_device()` but set the
`profile.nmdevice` directly. Also change the wait code to be OK to run
with `profile.profile` as None as in this case we don't care about which
profile is activating.

Also removed the check on `DEACTIVATING` as device might also in
`NM_DEVICE_STATE_DISCONNECTED` state, we should keep waiting as long as
device got `NEW_ACTIVATION`.

Signed-off-by: Gris Ge <fge@redhat.com>